+++ tests/core/emscripten_scan_registers.txt
@@ -1,1 +1,5 @@
-scan sees: 314159
+scan
+a 314159, 21828
+scan
+b 314160, 21827
+ok
